backend/foodgram/requirements.txt
asgiref==3.7.2
blosc2==2.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
defusedxml==0.8.0rc2
Django==4.2.7
django-extensions==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
exa==0.5.26
exceptiongroup==1.1.3
fonttools==4.44.3
idna==3.4
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.12.0
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.8.1
mpmath==1.3.0
msgpack==1.0.7
ndindex==1.7
networkx==3.2.1
numba==0.58.1
numexpr==2.8.7
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pluggy==1.3.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytils==0.4.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
seaborn==0.13.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
svgwrite==1.4.3
sympy==1.12
tables==3.9.1
tomli==2.0.1
Tree==0.2.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.2
zipp==3.17.0
